+++ app/navigation/ElderNavigator.tsx
@@ -3,7 +3,7 @@ import ElderTabNavigator from "./ElderTabNavigator";
 import SharedHealthStackNavigator from "./SharedHealthStackNavigator";
 import { ElderTabParamList } from "./navigation";
 import AssignCaregiverScreen from "../elder_only/assigncaregiver";
-import ElderPerimeterMap from "../elder_only/map_elder";
+import ElderPerimeterMap from "../elder_only/MapElder";
 import RemoveCaregiver from "../elder_only/removecaregiver";
 import ViewArduino from "../elder_only/viewarduino";
